chore: move container.tsx file to layouts folder and update import paths
